+++ Tenta/Generatable.swift
@@ -9,7 +9,7 @@ import Foundation
  Represents a type that has a default generator. Used alongside tests to make a shorthand for the default generator.
  */
 public protocol Generatable {
-    static var generator: Generator<Self> { get }
+    static var generator: AnyGenerator<Self> { get }
 }
 
 extension Generatable {