+++ Tenta/Generators/Character+Generatable.swift
@@ -5,31 +5,31 @@
 
 import Foundation
 
-public extension Generator where ValueToTest == Character {
-    static var char: Generator<Character> {
-        Generator<UInt8>.element(from: 32...255).map { uInt8 in
+public extension AnyGenerator where ValueToTest == Character {
+    static var char: AnyGenerator<Character> {
+        AnyGenerator<UInt8>.element(from: 32...255).map { uInt8 in
             Character(UnicodeScalar(uInt8))
         }
     }
 
-    static var utf8: Generator<Character> {
-        Generator<UInt32>.uInt32.compactMap { int in
+    static var utf8: AnyGenerator<Character> {
+        AnyGenerator<UInt32>.uInt32.compactMap { int in
             Unicode.Scalar(int).map { Character($0) }
         }
     }
 
-    static var alphaNumeric: Generator<Character> {
+    static var alphaNumeric: AnyGenerator<Character> {
         let alphabet = "abcdefghijklmnopqrstuvwxyz"
-        return Generator<Character>.element(from: alphabet + alphabet.uppercased() + "1234567890")
+        return AnyGenerator<Character>.element(from: alphabet + alphabet.uppercased() + "1234567890")
     }
 
-    func generateString() -> Generator<String> {
+    func generateString() -> AnyGenerator<String> {
         reduce("") { $0 + String($1) }
     }
 }
 
 extension Character: Generatable {
-    public static var generator: Generator<Character> {
-        Generator<Character>.char
+    public static var generator: AnyGenerator<Character> {
+        AnyGenerator<Character>.char
     }
 }