+++ Tenta/Generators/Data+Generatable.swift
@@ -5,20 +5,20 @@
 
 import Foundation
 
-public extension Generator where ValueToTest == Data {
-    static var data: Generator<Data> {
+public extension AnyGenerator where ValueToTest == Data {
+    static var data: AnyGenerator<Data> {
         [UInt8].generator.map { Data($0) }
     }
 }
 
 extension Data: Generatable {
-    public static var generator: Generator<Data> {
-        Tenta.Generator<Data>.data
+    public static var generator: AnyGenerator<Data> {
+        Tenta.AnyGenerator<Data>.data
     }
 }
 
-public extension Generator where ValueToTest == URLQueryItem {
-    static var urlQueryItem: Generator<URLQueryItem> {
+public extension AnyGenerator where ValueToTest == URLQueryItem {
+    static var urlQueryItem: AnyGenerator<URLQueryItem> {
         String.generator.combine(with: (String?).generator).map {
             URLQueryItem(name: $0, value: $1)
         }
@@ -26,30 +26,30 @@ public extension Generator where ValueToTest == URLQueryItem {
 }
 
 extension URLQueryItem: Generatable {
-    public static var generator: Generator<URLQueryItem> {
-        Tenta.Generator<URLQueryItem>.urlQueryItem
+    public static var generator: AnyGenerator<URLQueryItem> {
+        Tenta.AnyGenerator<URLQueryItem>.urlQueryItem
     }
 }
 
-//public extension Generator where ValueToTest == URLComponents {
-//    static var urlComponents: Generator<URLComponents> {
-//        return Generator.simple { (constructor: inout Constructor) -> URLComponents in
+//public extension AnyGenerator where ValueToTest == URLComponents {
+//    static var urlComponents: AnyGenerator<URLComponents> {
+//        return AnyGenerator.simple { (constructor: inout Constructor) -> URLComponents in
 //            var components = URLComponents()
-//            components.fragment = Generator<String>.alphaNumeric.optional().generateUsing(&constructor)
-//            components.host = Generator<String>.alphaNumeric.optional().generateUsing(&constructor)
-//            components.password = Generator<String>.alphaNumeric.optional().generateUsing(&constructor)
-//            components.path = Generator<String>.alphaNumeric.generateUsing(&constructor)
+//            components.fragment = AnyGenerator<String>.alphaNumeric.optional().generateUsing(&constructor)
+//            components.host = AnyGenerator<String>.alphaNumeric.optional().generateUsing(&constructor)
+//            components.password = AnyGenerator<String>.alphaNumeric.optional().generateUsing(&constructor)
+//            components.path = AnyGenerator<String>.alphaNumeric.generateUsing(&constructor)
 //            components.port = Int.generator.map(abs).generateUsing(&constructor)
 //            components.queryItems = [URLQueryItem].generator.optional().generateUsing(&constructor)
-//            components.scheme = Generator<String>.alphaNumeric.optional().generateUsing(&constructor)
-//            components.user = Generator<String>.alphaNumeric.optional().generateUsing(&constructor)
+//            components.scheme = AnyGenerator<String>.alphaNumeric.optional().generateUsing(&constructor)
+//            components.user = AnyGenerator<String>.alphaNumeric.optional().generateUsing(&constructor)
 //            return components
 //        }
 //    }
 //}
 //
 //extension URLComponents: Generatable {
-//    public static var generator: Generator<URLComponents> {
+//    public static var generator: AnyGenerator<URLComponents> {
 //        return Tenta.Generator<URLComponents>.urlComponents
 //    }
 //}